修复部分造成应用crash的bug，增加数据更新入口

+++ app/src/main/java/model/recordLogger/Role.java
@@ -18,7 +18,7 @@ public class Role implements Serializable{
 
     public class RoleInfo implements Serializable{
         public String RoleName;
-        public int RoleID;
+        public long RoleID;
         public int RoleLevel;
         public int JumpValue;
         public int WinCount;
